Patch the initial-data-slurp logic in init.f.
Believe it or not, the "long.dat" etc. files are actually in the
correct format already! Each line is intended to be slurped into a
string variable named `FMT` and then passed to `WRITE(6,FMT)` —
basically the equivalent of `printf(s)` in C. The parentheses and
single-quotes on each line are part of the format string!