Modernize the random number generator.
Eliminate the unknown dependencies `TOD` ("time of day")
and `RDMIN` (seed the PRNG). Implement `RDM` to call the
now-standard `RANDOM_NUMBER` library function instead.

Thanks to @welash for pointing out that RDM should be a
`REAL FUNCTION` instead of an `INTEGER FUNCTION` (although
surprisingly it seems to work the same either way).