Fix runtime errors with the "HUH" message.
When you hit enter without giving any command,
the game tries to print the message HUH as a format string...
but it's not a valid format string!

    Fortran runtime error: Missing initial left parenthesis in format

Reuse the existing format string 1001 for this message.

Secondly, heasm66 reports that `REWIND 5` crashes the game
on Cygwin (but not on Linux).

Thirdly, the lack of carriage-control formatting on the message
makes it print weirdly with `asa.py`, so let's fix that too.

This code may have worked fine on VS Fortran back in the day,
for all I know.