Save XLIM in WWOLF and GGNOME.
Thanks to eriktorbjorn for identifying this issue in
https://github.com/Quuxplusone/Castlequest/issues/8 !
It seems that VS FORTRAN, like many old Fortrans, placed all
subroutine-local variables into static storage. The manual
says:

> Because VS FORTRAN saves these names without user action,
> the `SAVE` statement serves only as a documentation aid.

http://bitsavers.trailing-edge.com/pdf/ibm/370/OS_VS/fortran/GC26-3986-0_VS_FORTRAN_Prog_Ref_Feb81.pdf

However, modern Fortrans use stack storage for local variables,
so when `WOLF .EQ. .TRUE.`, we'd be using `XLIM` uninitialized,
which seems to be undefined behavior according to modern Fortrans.
The solution is to add `SAVE XLIM` at the top of the subroutine.

After this patch, rather than using "uninitialized" probabilities,
the game will correctly give the werewolf a 7.5% chance of success
on its first attack, 22.5% on its second attack, and so on, up to
a 100% chance of killing the player within eight attacks.

The gnome gets a similar fix, with 7.5% on the first attack,
27.5% on the second attack, and so on, up to a 100% chance of
killing the player within six attacks.